requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
